Migrated add component functionality to separate file
